requirements.txt
streamlit
fpdf
pypdf
faiss-cpu
langchain
langchain_community
sentence-transformers
openai
euriai
PyPDF2
python-docx
python-pptx
openpyxl
pandas
lxml